requirements.txt
aiogram
python-dotenv==0.19.0
